Blacklist fully implemented as a function of the AuthenticationService.cs. Now each endpoint verifies if the JWT token is valid before performing any operation. If the token is part of the Blacklist table, it will be return an HTTP 401 Unauthorized response. This has been fully applied to all the endpoints already created that are only accessible to Admin users.
Products Model was updated to include the Description and Category fields and a migration was done to push these changes to the DB.

+++ Endpoints/ProductEndpoints.cs
@@ -1,5 +1,6 @@
 ﻿using Project.Models;
 using Project.Services.Repository;
+using Project.Services;
 
 
 namespace Project.Endpoints
@@ -25,16 +26,28 @@ namespace Project.Endpoints
                 return product != null ? Results.Ok(product) : Results.NotFound($"There is no product with ID {productId}.");                
             });
 
-            app.MapPost("/Product", async (IRepository<Product> productRepo, Product newProduct) =>
+            app.MapPost("/Product", async (IRepository<Product> productRepo, Product newProduct, IAuthenticationService auth, HttpRequest request) =>
             {
-                return await productRepo.AddAsync(newProduct);
-            });
+                bool tokenIsValid = await auth.ValidateJwtToken(request);
+                if (!tokenIsValid) return Results.Unauthorized();
 
-            app.MapDelete("/Product/{ProductId}", async (IRepository<Product> productRepo, int productId) =>
+                await productRepo.AddAsync(newProduct);
+
+                return Results.Created($"/Product/{newProduct.ProductId}", newProduct);
+
+            }).RequireAuthorization("adminAccess");
+
+            app.MapDelete("/Product/{ProductId}", async (IRepository<Product> productRepo, int productId, IAuthenticationService auth, HttpRequest request) =>
             {
+                bool tokenIsValid = await auth.ValidateJwtToken(request);
+                if (!tokenIsValid) return Results.Unauthorized();
+
                 Product product = await productRepo.GetByIdAsync(productId);
-                return await productRepo.DeleteAsync(product); 
-            });
+                await productRepo.DeleteAsync(product);
+                
+                return Results.NoContent();
+
+            }).RequireAuthorization("adminAccess");
         }
     }
 }
